Major Change - Made most of PK from int to Long. Additional remarks -
1. Bug fixed - For every offered ride, it was matching with only one
ride request irrespective of more seats available. So now this has been
changed and this will match all ride request till all seats are filled
up.

2. TODO later, we are not able to store ride id / ride request id as
long in mongo, as by default it uses Integer and we need to provide more
customization in Json while creating records to store in Long format.
Though this doesn't have any impact immediately, so left it for the time
being.

+++ RSBillingSystem/src/com/digitusrevolution/rideshare/billing/business/AccountBusinessService.java
@@ -16,7 +16,7 @@ public class AccountBusinessService {
 
 	private static final Logger logger = LogManager.getLogger(AccountBusinessService.class.getName());
 
-	public void addMoneyToWallet(int accountNumber, float amount) {
+	public void addMoneyToWallet(long accountNumber, float amount) {
 		Session session = HibernateUtil.getSessionFactory().getCurrentSession();
 		Transaction transaction = null;
 		try {
@@ -41,7 +41,7 @@ public class AccountBusinessService {
 		}
 	}
 	
-	public void redeemFromWallet(int accountNumber, float amount) {
+	public void redeemFromWallet(long accountNumber, float amount) {
 		Session session = HibernateUtil.getSessionFactory().getCurrentSession();
 		Transaction transaction = null;
 		try {
@@ -66,7 +66,7 @@ public class AccountBusinessService {
 		}
 	}
 	
-	public List<com.digitusrevolution.rideshare.model.billing.domain.core.Transaction> getTransactions(int accountNumber, int page) {
+	public List<com.digitusrevolution.rideshare.model.billing.domain.core.Transaction> getTransactions(long accountNumber, int page) {
 		Session session = HibernateUtil.getSessionFactory().getCurrentSession();
 		Transaction transaction = null;
 		List<com.digitusrevolution.rideshare.model.billing.domain.core.Transaction> transactions = new ArrayList<>();
