+++ RSCommon/src/com/digitusrevolution/rideshare/common/inf/AuthServiceInf.java
@@ -2,7 +2,7 @@ package com.digitusrevolution.rideshare.common.inf;
 
 public interface AuthServiceInf {
 
-	public String getToken(int userId);
+	public String getToken(long userId);
 	public boolean verifyToken(String token);
 	
 }
