+++ RSModel/src/com/digitusrevolution/rideshare/model/user/data/PhotoEntity.java
@@ -11,16 +11,15 @@ import javax.persistence.Table;
 public class PhotoEntity {
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private int id;
+	private long id;
 	private String imageLocation;
 	
-	public int getId() {
+	public long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
-	
 	public String getImageLocation() {
 		return imageLocation;
 	}
